+++ lib/src/main/java/module-info.java
@@ -5,6 +5,7 @@ module org.seuffert.panvpx {
     exports org.seuffert.panvpx;
     exports org.seuffert.panvpx.core;
     exports org.seuffert.panvpx.vp8;
+    exports org.seuffert.panvpx.vp9;
 
     // Do NOT export org.seuffert.panvpx.ffi to keep jextract bindings internal
 }